Remove randr dependeency in VarInfo
