+++ src/core/varinfo.jl
@@ -263,36 +263,36 @@ updategid!(vi, vn, spl) = begin
   end
 end
 
-# This method is called when sampler is missing
-# NOTE: this used for initialize VarInfo, i.e. vi = model()
-# NOTE: this method is also used by IS
-randr(vi::VarInfo, vn::VarName, dist::Distribution) = randr(vi, vn, dist, false)
-randr(vi::VarInfo, vn::VarName, dist::Distribution, count::Bool) = begin
-  vi.index = count ? vi.index + 1 : vi.index
-  if ~haskey(vi, vn)
-    nwevar!(vi, vn, dist)
-  else
-    if count checkindex(vn, vi) end
-    r = vi[vn]
-    vi.logp += logpdf(dist, r, istransformed(vi, vn))
-    r
-  end
-end
-
-# Random with replaying
-# NOTE: this method is used by PG
-randr(vi::VarInfo, vn::VarName, dist::Distribution, spl::Sampler) = randr(vi, vn, dist, spl, false)
-randr(vi::VarInfo, vn::VarName, dist::Distribution, spl::Sampler, count::Bool) = begin
-  vi.index = count ? vi.index + 1 : vi.index
-  if ~haskey(vi, vn)
-    nwevar!(vi, vn, dist, spl.alg.gid)
-  elseif isnan(vi, vn)
-    r = rand(dist)
-    setval!(vi, vectorize(dist, r), vn)
-    r
-  else
-    if count checkindex(vn, vi, spl) end
-    updategid!(vi, vn, spl)
-    vi[vn]
-  end
-end
+# # This method is called when sampler is missing
+# # NOTE: this used for initialize VarInfo, i.e. vi = model()
+# # NOTE: this method is also used by IS
+# randr(vi::VarInfo, vn::VarName, dist::Distribution) = randr(vi, vn, dist, false)
+# randr(vi::VarInfo, vn::VarName, dist::Distribution, count::Bool) = begin
+#   vi.index = count ? vi.index + 1 : vi.index
+#   if ~haskey(vi, vn)
+#     nwevar!(vi, vn, dist)
+#   else
+#     if count checkindex(vn, vi) end
+#     r = vi[vn]
+#     vi.logp += logpdf(dist, r, istransformed(vi, vn))
+#     r
+#   end
+# end
+#
+# # Random with replaying
+# # NOTE: this method is used by PG
+# randr(vi::VarInfo, vn::VarName, dist::Distribution, spl::Sampler) = randr(vi, vn, dist, spl, false)
+# randr(vi::VarInfo, vn::VarName, dist::Distribution, spl::Sampler, count::Bool) = begin
+#   vi.index = count ? vi.index + 1 : vi.index
+#   if ~haskey(vi, vn)
+#     nwevar!(vi, vn, dist, spl.alg.gid)
+#   elseif isnan(vi, vn)
+#     r = rand(dist)
+#     setval!(vi, vectorize(dist, r), vn)
+#     r
+#   else
+#     if count checkindex(vn, vi, spl) end
+#     updategid!(vi, vn, spl)
+#     vi[vn]
+#   end
+# end
